requirements.txt
numpy==2.0.2
opencv-python-headless==4.10.0.84
uvicorn==0.31.1
fastapi==0.115.2
python-socketio==5.11.4
starlette~=0.40.0
face-recognition~=1.3.0